Extract AddDatePolicyFactory to public class. Fixes #391.
